// Qanat.Web/src/app/shared/generated/model/monitoring-well-data-dto.ts
export class MonitoringWellDataDto { 
    MonitoringWellID?: number;
    GeographyID?: number;
    MonitoringWellName?: string;
    SiteCode?: string;
    MonitoringWellSourceTypeDisplayName?: string;
    Latitude?: number;
    Longitude?: number;
    NumberOfMeasurements?: number;
    EarliestMeasurementDate?: string;
    LastMeasurementDate?: string;
    EarliestMeasurement?: number;
    LastMeasurement?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}